requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
dj-database-url==1.0.0
dj-static==0.0.6
Django==4.1.3
django-filter==22.1
django-stdimage==6.0.1
django-widget-tweaks==1.4.12
gunicorn==20.1.0
Pillow==9.2.0
psycopg2-binary==2.8.6
pytz==2022.2.1
sqlparse==0.4.2
static3==0.7.0
whitenoise==6.2.0
